+++ quantify_scheduler/backends/qblox/operation_handling/factory.py
@@ -17,7 +17,6 @@ from quantify_scheduler.enums import BinMode
 
 def get_operation_strategy(
     operation_info: OpInfo,
-    instruction_generated_pulses_enabled: bool,
     io_mode: IoMode,
 ) -> base.IOperationStrategy:
     """
@@ -27,9 +26,6 @@ def get_operation_strategy(
     ----------
     operation_info
         The operation we are building the strategy for.
-    instruction_generated_pulses_enabled
-        Specifies if instruction generated pulses (e.g. staircase through offsets) are
-        allowed. If set to False, only generically treated pulses are allowed.
     io_mode
         Either :attr:`.IoMode.REAL`, :attr:`.IoMode.IMAG` or :attr:`.IoMode.COMPLEX`
         depending on whether the signal affects only sequencer path0, path1 or both.
@@ -44,7 +40,6 @@ def get_operation_strategy(
 
     return _get_pulse_strategy(
         operation_info=operation_info,
-        instruction_generated_pulses_enabled=instruction_generated_pulses_enabled,
         io_mode=io_mode,
     )
 
@@ -77,7 +72,6 @@ def _get_acquisition_strategy(
 
 def _get_pulse_strategy(
     operation_info: OpInfo,
-    instruction_generated_pulses_enabled: bool,
     io_mode: IoMode,
 ) -> base.IOperationStrategy:
     """Handles the logic for determining the correct pulse type."""
@@ -99,15 +93,6 @@ def _get_pulse_strategy(
             return virtual.NcoSetClockFrequencyStrategy(operation_info)
         else:
             return virtual.IdleStrategy(operation_info)
-
-    elif instruction_generated_pulses_enabled:
-        wf_func = operation_info.data["wf_func"]
-
-        if wf_func == "quantify_scheduler.waveforms.square":
-            return pulses.StitchedSquarePulseStrategy(operation_info, io_mode)
-
-        elif wf_func == "quantify_scheduler.waveforms.staircase":
-            return pulses.StaircasePulseStrategy(operation_info, io_mode)
 
     elif operation_info.name == "MarkerPulse":
         return pulses.MarkerPulseStrategy(operation_info, io_mode)